feat(apply): The "apply" now takes "runtime" as toplevel argument
Capabilities and Generators will still look for the `runtime` value
in their properties, but it will be assumed that the value will be
the same for all of them within a single application (which is a
single call to `apply()`. This is more along the line of what the
REST endpoint will expect.