feat(deploy): Added push functionality
New feature that can push local code to OpenShift and start a build.
This can either be source code which will trigger a build of necessary
or a binary can be pushed that will be run directly.
A new script named "push" is added to do this from the CLI.

Fixes #7 Enable inner-loop development